Feature #50611961: Make URL detection work
To do this, we hack UIBubbleTableView and add support for UITextViews.
Some magic needs to occur to handle the case of a single line of text
since it needs to be sized to minwidth. For now, hack a hardcoded value.

This will be cleaned up and pull-requested back to UIBubbleTableView...