Feature #50794261: insta-load conversations
The basic design here is twofold

1. Maintain a list of queued messages, which are real NSManagedObjects*,
but have not been sent to the server yet; work on this queue in FIFO order.

2. Create fake gabs with the gab_id negative.

When the first message returns with a real gab_id, sync will create the real
gab and update the original message (search by code). Then, update all
queued messages with the real gab_id, and delete the fake gab.

Additionally, refactor the API helper code so that there are nonblocking
and blocking API endpoint calls.

Remove QUIET option; handle failure (likely internet not available) globally.
(Coming in a later commit)