Fix frozen login
The basic problem was that login with an authenticated token had a race
condition against didRegisterDeviceToken...

To deal with this, refactor the login code.

1. Fire a NSNotification when login is complete; we listen on this
to fire fetchUserData - which ultimately calls postLogin which
clears the login screen.

2. When doing an quick-cached-login, don't yet fireLogin (we need
to refactor the fetchUserData first). Just go straight to postLogin.

3. Always cache the provider when we cache the login token. This
means we silently login in with the right provider.

4. Fire a notification when we get the device token, and try to
relogin. If we didn't login successfully even though we had an
access_token - e.g. we did a quick-cached-login and we beat
didRegisterDeviceToken - this will log us in properly.

There is a lot of further refactoring work TBD here.