PT #55613776: make it so we instantly update gabs when we write
This fixes the case when you write a message, go back to the
main gab, and it takes a few secs to update (especially on slow
internet) because we have to query the list of gabs to get the
content_summary and updated_at times. This requires an optional
server side change but we gracefully degrade if we don't get it.

What we specifically do is try to get an additional JSON hash
"gab", and if we get it, parse it using the regular logic.

However, we first set the updated_at locally to the created_at
date of the message we just received; this makes sure we
correctly update the needs_update iff we REALLY need an update

Note that this doesn't handle the case where you could get an
APN while parsing the response...