PT #55744166: update the main gab view due to a GabUpdate
However, for perf's sake, only do it if the main gab is visible.
When teh main gab view comes back we update the table in willAppear:

The main case to fix here was when an message APN arrives on the
main gab; we sent a gabUpdate but since main gab wasn't listening,
there was no refresh.