PT #55759772: fix first load of maingabview
Because we were making the view visible and ready before we
figured out if we were logged in via cached token or whatever,
we were loading cached data from a blank store. this meant that
the user didn't see the cached local data until the first full
pull from the server.