PT #55763754: fix up the login controller
Refactor login controller to be a first class controller that is
always pushed on first.

Refactor main gab view controller to be pushed on as needed after
login. This means we get correct behavior on first table reload
since the managed context is always guarenteed to be set.

Unpush the main gab view controller on logout.

* When opening an APN from death, have no animations straight
to the gab view.

* When opening an APN from backgrounded, have no animations
straight to the gab view.

* When opening an APN from foreground, do nothing btu refresh
screens as needed.

* When cached login start, go straight to the main screen with
no animation.

* When login with cached credentials, throw up login screen
with no buttons and then login.

* Otherwise, throw up login screen with buttons with button
animation.

* When logging out, throw up login screen with no button
animation.